+++ scrapyd/website.py
@@ -10,12 +10,18 @@ from scrapyd.interfaces import IEggStorage, IPoller, ISpiderScheduler
 from scrapyd.jobstorage import job_items_url, job_log_url
 
 
+class PrefixHeaderMixin:
+    def get_base_path(self, txrequest):
+        return txrequest.getHeader(self.prefix_header) or ''
+
+
 class Root(resource.Resource):
 
     def __init__(self, config, app):
         resource.Resource.__init__(self)
         self.debug = config.getboolean('debug', False)
         self.runner = config.get('runner')
+        self.prefix_header = config.get('prefix_header')
         logsdir = config.get('logs_dir')
         itemsdir = config.get('items_dir')
         self.local_items = itemsdir and (urlparse(itemsdir).scheme.lower() in ['', 'file'])
@@ -55,16 +61,18 @@ class Root(resource.Resource):
         return self.app.getComponent(IPoller)
 
 
-class Home(resource.Resource):
+class Home(PrefixHeaderMixin, resource.Resource):
 
     def __init__(self, root, local_items):
         resource.Resource.__init__(self)
         self.root = root
         self.local_items = local_items
+        self.prefix_header = root.prefix_header
 
     def render_GET(self, txrequest):
         vars = {
-            'projects': ', '.join(self.root.scheduler.list_projects())
+            'projects': ', '.join(self.root.scheduler.list_projects()),
+            'base_path': self.get_base_path(txrequest),
         }
         s = """
 <html>
@@ -73,12 +81,12 @@ class Home(resource.Resource):
 <h1>Scrapyd</h1>
 <p>Available projects: <b>%(projects)s</b></p>
 <ul>
-<li><a href="/jobs">Jobs</a></li>
-""" % vars
+<li><a href="%(base_path)s/jobs">Jobs</a></li>
+"""
         if self.local_items:
-            s += '<li><a href="/items/">Items</a></li>'
+            s += '<li><a href="%(base_path)s/items/">Items</a></li>'
         s += """
-<li><a href="/logs/">Logs</a></li>
+<li><a href="%(base_path)s/logs/">Logs</a></li>
 <li><a href="http://scrapyd.readthedocs.org/en/latest/">Documentation</a></li>
 </ul>
 
@@ -94,9 +102,9 @@ monitoring)</p>
 <a href="http://scrapyd.readthedocs.org/en/latest/">Scrapyd documentation</a></p>
 </body>
 </html>
-""" % vars
+"""
         txrequest.setHeader('Content-Type', 'text/html; charset=utf-8')
-        s = s.encode('utf8')
+        s = (s % vars).encode('utf8')
         txrequest.setHeader('Content-Length', str(len(s)))
         return s
 
@@ -109,15 +117,16 @@ def microsec_trunc(timelike):
     return timelike - timedelta(microseconds=ms)
 
 
-class Jobs(resource.Resource):
+class Jobs(PrefixHeaderMixin, resource.Resource):
 
     def __init__(self, root, local_items):
         resource.Resource.__init__(self)
         self.root = root
         self.local_items = local_items
+        self.prefix_header = root.prefix_header
 
     cancel_button = """
-    <form method="post" action="/cancel.json">
+    <form method="post" action="{base_path}/cancel.json">
     <input type="hidden" name="project" value="{project}"/>
     <input type="hidden" name="job" value="{jobid}"/>
     <input type="submit" style="float: left;" value="Cancel"/>
@@ -161,7 +170,7 @@ class Jobs(resource.Resource):
             '<style type="text/css">' + self.gen_css() + '</style>'
             '</head>'
             '<body><h1>Jobs</h1>'
-            '<p><a href="..">Go up</a></p>'
+            '<p><a href="./">Go up</a></p>'
             + self.prep_table() +
             '</body>'
             '</html>'
@@ -192,7 +201,7 @@ class Jobs(resource.Resource):
                 "Project": project,
                 "Spider": m['name'],
                 "Job": m['_job'],
-                "Cancel": self.cancel_button(project=project, jobid=m['_job']),
+                "Cancel": self.cancel_button(project=project, jobid=m['_job'], base_path=self.base_path),
             })
             for project, queue in self.root.poller.queues.items()
             for m in queue.list()
@@ -207,9 +216,9 @@ class Jobs(resource.Resource):
                 "PID": p.pid,
                 "Start": microsec_trunc(p.start_time),
                 "Runtime": microsec_trunc(datetime.now() - p.start_time),
-                "Log": f'<a href="{job_log_url(p)}">Log</a>',
-                "Items": f'<a href="{job_items_url(p)}">Items</a>',
-                "Cancel": self.cancel_button(project=p.project, jobid=p.job),
+                "Log": f'<a href="{self.base_path}{job_log_url(p)}">Log</a>',
+                "Items": f'<a href="{self.base_path}{job_items_url(p)}">Items</a>',
+                "Cancel": self.cancel_button(project=p.project, jobid=p.job, base_path=self.base_path),
             })
             for p in self.root.launcher.processes.values()
         )
@@ -223,13 +232,14 @@ class Jobs(resource.Resource):
                 "Start": microsec_trunc(p.start_time),
                 "Runtime": microsec_trunc(p.end_time - p.start_time),
                 "Finish": microsec_trunc(p.end_time),
-                "Log": f'<a href="{job_log_url(p)}">Log</a>',
-                "Items": f'<a href="{job_items_url(p)}">Items</a>',
+                "Log": f'<a href="{self.base_path}{job_log_url(p)}">Log</a>',
+                "Items": f'<a href="{self.base_path}{job_items_url(p)}">Items</a>',
             })
             for p in self.root.launcher.finished
         )
 
     def render(self, txrequest):
+        self.base_path = self.get_base_path(txrequest)
         doc = self.prep_doc()
         txrequest.setHeader('Content-Type', 'text/html; charset=utf-8')
         doc = doc.encode('utf-8')